+++ src/SeablastConfiguration.php
@@ -5,12 +5,16 @@ declare(strict_types=1);
 namespace Seablast\Seablast;
 
 use Seablast\Seablast\SeablastConfigurationException;
+use Seablast\Seablast\SeablastMysqli;
+use Tracy\Debugger;
 use Webmozart\Assert\Assert;
 
 class SeablastConfiguration
 {
     use \Nette\SmartObject;
 
+    /** @var SeablastMysqli */
+    private $connection = null;
     /** @var SeablastFlag */
     public $flag;
     /** @var array<array<string[]>> */
@@ -27,6 +31,55 @@ class SeablastConfiguration
     public function __construct()
     {
         $this->flag = new SeablastFlag();
+    }
+
+    /**
+     * Access to database with lazy initialisation
+     * @return SeablastMysqli
+     */
+    public function dbms(): SeablastMysqli
+    {
+        //Lazy initialisation
+        if (is_null($this->connection)) {
+            Debugger::barDump('Creating database connection');
+            $this->dbmsCreate();
+        }
+        return $this->connection;
+    }
+
+    /**
+     * Creates a database connection and sets up charset.
+     * @return void
+     */
+    private function dbmsCreate(): void
+    {
+        $phinx = $this->dbmsReadPhinx();
+        // todo Assert:: environment dle SB_phinx or default environment ... Parametry foreach Assert:: string
+        $environment = 'developmnent'; // todo config ?? $phinx['environments']['default_environment']
+        $this->connection = new SeablastMysqli(
+            $phinx['environments'][$environment]['host'], // todo fix localhost
+            $phinx['environments'][$environment]['user'],
+            $phinx['environments'][$environment]['pass'],
+            $phinx['environments'][$environment]['name'],
+            $phinx['environments'][$environment]['port'] ?? null
+        );
+        // todo does this really differentiate between successful connection, failed connection and no connection?
+        Assert::isAOf($this->connection, '\Seablast\Seablast\SeablastMysqli');
+        $this->connection->set_charset('utf8'); // TODO viz configuration
+    }
+
+    /**
+     * TODO ignore: `Method Seablast\Seablast\SeablastConfiguration::dbmsReadPhinx() should return array but
+     * TODO ignore: return statement is missing.`
+     * @return array<mixed>
+     * @throws \Exception
+     */
+    private static function dbmsReadPhinx(): array
+    {
+        if (!file_exists(APP_DIR . '/conf/phinx.local.php')) {
+            throw new \Exception('Give credentials to use database');
+        }
+        include APP_DIR . '/conf/phinx.local.php'; // which contains the return statement
     }
 
     /**
@@ -200,11 +253,11 @@ class SeablastConfiguration
 
     /**
      * Debug
-     * @return void
+     * @ return void
      */
-    public function dump(): void
-    {
-        var_dump($this->optionsBool);
-        var_dump($this->optionsInt, $this->optionsString);
-    }
+//    public function dump(): void
+//    {
+//        var_dump($this->optionsBool);
+//        var_dump($this->optionsInt, $this->optionsString);
+//    }
 }