+++ src/Superglobals.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Seablast\Seablast;
 
-//use Webmozart\Assert\Assert;
-
 class Superglobals
 {
     use \Nette\SmartObject;
@@ -28,7 +26,6 @@ class Superglobals
      */
     public function __construct(array $get = [], array $post = [], array $server = [], array $session = [])
     {
-        //todo named parameters? one array?
         $this->get = $get;
         $this->post = $post;
         $this->server = $server;